+++ tests/unit/test_client.py
@@ -1,19 +1,20 @@
 from finetuner.constants import (
     API_VERSION,
     CONFIG,
-    CPUS,
     DELETE,
     EXAMPLES,
     EXPERIMENTS,
     GET,
-    GPUS,
     LOGS,
     METRICS,
     NAME,
     POST,
     RUNS,
     STATUS,
+    SYNTHESIS_TASK,
+    TRAINING_TASK,
 )
+from finetuner.data import DATA_SYNTHESIS_EN
 from finetuner.experiment import Experiment
 
 
@@ -70,7 +71,7 @@ def test_delete_experiments(client_mocker):
     assert sent_request['method'] == DELETE
 
 
-def test_create_run(client_mocker, experiment_name='exp', run_name='run'):
+def test_create_training_run(client_mocker, experiment_name='exp', run_name='run'):
     config = Experiment._create_finetuning_config(
         model='resnet50',
         train_data='data name',
@@ -81,6 +82,7 @@ def test_create_run(client_mocker, experiment_name='exp', run_name='run'):
         experiment_name=experiment_name,
         run_name=run_name,
         run_config=config,
+        task=TRAINING_TASK,
         device='cpu',
         cpus=1,
         gpus=1,
@@ -91,8 +93,32 @@ def test_create_run(client_mocker, experiment_name='exp', run_name='run'):
     assert sent_request['method'] == POST
     assert sent_request['json_data'][NAME] == run_name
     assert sent_request['json_data'][CONFIG] == config
-    assert sent_request['json_data'][CPUS] == 1
-    assert sent_request['json_data'][GPUS] == 1
+
+
+def test_create_synthesis_run(client_mocker, experiment_name='exp', run_name='run'):
+    config = Experiment._create_synthesis_config(
+        query_data='query_data_name',
+        corpus_data='corpus_data_name',
+        models=DATA_SYNTHESIS_EN,
+        num_relations=3,
+        experiment_name=experiment_name,
+        run_name=run_name,
+    )
+    sent_request = client_mocker.create_run(
+        experiment_name=experiment_name,
+        run_name=run_name,
+        run_config=config,
+        task=SYNTHESIS_TASK,
+        device='cpu',
+        cpus=1,
+        gpus=1,
+    )
+    assert sent_request['url'] == client_mocker._construct_url(
+        client_mocker._base_url, API_VERSION, EXPERIMENTS, experiment_name, RUNS
+    )
+    assert sent_request['method'] == POST
+    assert sent_request['json_data'][NAME] == run_name
+    assert sent_request['json_data'][CONFIG] == config
 
 
 def test_get_run(client_mocker, experiment_name='exp', run_name='run1'):